PR feedback: move statistics utils
