+++ frontend/src/utils/StatisticsUtils.ts
